third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/margin-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'margin' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
Harness: the test ran to completion.
